pyproject.toml
[project]
name = "pylangdb"
version = "0.2.6"
description = "Comprehensive Google ADK integration package for LangDB with enhanced agents, LLM integration, and distributed tracing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api>=1.18.0",
    "opentelemetry-sdk>=1.18.0",
    "opentelemetry-exporter-otlp>=1.18.0",
    "uuid>=1.0.0"
]

[project.optional-dependencies]
# Client library extras
adk = [
    "google-adk>=1.3.0",
]
agno = [
    "agno>=1.7.0",
    "openinference-instrumentation-agno>=0.1.14",
]
crewai = [
    "openinference-instrumentation-crewai>=0.1.14",
    "litellm>=1.73.2",
]
langchain = [
    "openinference-instrumentation-langchain>=0.1.14",
]
openai = [
    "openai>=1.60.2",
    "opentelemetry-instrumentation-openai-agents>=0.1.14",
]
client = [
    "pandas>=2.2.2",
    "requests>=2.32.3",
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
]
# All dependencies
all = [
    "agno>=1.7.0",
    "google-adk>=1.3.0",
    "litellm>=1.73.2",
    "openai>=1.60.2",
    "opentelemetry-instrumentation-openai-agents>=0.1.14",
    "openinference-instrumentation-agno>=0.1.14",
    "openinference-instrumentation-crewai>=0.1.14",
    "openinference-instrumentation-langchain>=0.1.14",
]
